# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_nate_go-boilerplate_pkg_response.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
...
  internal_http_handler.UpdateProfileRequest:
    properties:
      email:
        example: newemail@example.com
        type: string
      phone:
        example: "+1987654321"
        type: string
    type: object
  internal_http_handler.UserListResponse:
    properties:
      page:
        example: 1
        type: integer
      per_page:
        example: 20
        type: integer
      total:
        example: 42
        type: integer
      total_pages:
        example: 3
        type: integer
      users:
        items:
          $ref: '#/definitions/internal_http_handler.UserResponse'
        type: array
    type: object
  internal_http_handler.UserResponse:
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      phone:
        example: "+1234567890"
        type: string
      role:
        example: user
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/nate/go-boilerplate
  description: |-
    A production-ready Go REST API boilerplate with authentication, notes, and user management.
    Features include JWT authentication, rate limiting, soft deletes, and Redis-based token storage.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all users. Admin access required.
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/internal_http_handler.UserListResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a user by ID. Admin access required.
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User deleted successfully
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a specific user's profile. Admin access required.
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/internal_http_handler.UserResponse'
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - users
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/internal_http_handler.UserResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's email or phone. At least one of
        email or phone must be provided.
      parameters:
      - description: Profile updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user profile
          schema:
            $ref: '#/definitions/internal_http_handler.UserResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "409":
          description: Email or phone already in use
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_nate_go-boilerplate_pkg_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"